MOVE: `xdg_server_decoration` handler to sep. file
