+++ src/compositor/shell/xdg/handlers.rs
@@ -61,71 +61,3 @@ impl<BEnd: Backend> XdgActivationHandler for Navda<BEnd> {
 }
 delegate_xdg_activation!(@<BEnd: Backend + 'static> Navda<BEnd>);
 
-impl<BEnd: Backend> XdgDecorationHandler for Navda<BEnd> {
-    fn new_decoration(&mut self, toplevel: ToplevelSurface) {
-        toplevel.with_pending_state(|state| {
-            state.decoration_mode = Some(Mode::ClientSide);
-        });
-        toplevel.send_configure();
-    }
-    fn request_mode(&mut self, toplevel: ToplevelSurface, mode: Mode) {
-        if let Some(w) = self
-            .space
-            .elements()
-            .find(|window| matches!(window, AvWindow::Wayland(w) if w.toplevel() == &toplevel))
-        {
-            toplevel.with_pending_state(|state| {
-                state.decoration_mode = Some(match mode {
-                    Mode::ServerSide => {
-                        // TODO(Sammy99jsp) Server-side decoration.
-                        // w.set_ssd(true);
-                        Mode::ServerSide
-                    }
-                    _ => {
-                        // TODO(Sammy99jsp) Server-side decoration.
-                        // w.set_ssd(false);
-                        Mode::ClientSide
-                    }
-                });
-            });
-
-            let initial_configure_sent = with_states(toplevel.wl_surface(), |states| {
-                states
-                    .data_map
-                    .get::<XdgToplevelSurfaceData>()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .initial_configure_sent
-            });
-            if initial_configure_sent {
-                toplevel.send_configure();
-            }
-        }
-    }
-    fn unset_mode(&mut self, toplevel: ToplevelSurface) {
-        if let Some(w) = self
-            .space
-            .elements()
-            .find(|window| matches!(window, AvWindow::Wayland(w) if w.toplevel() == &toplevel))
-        {
-            // w.set_ssd(false);
-            toplevel.with_pending_state(|state| {
-                state.decoration_mode = Some(Mode::ClientSide);
-            });
-            let initial_configure_sent = with_states(toplevel.wl_surface(), |states| {
-                states
-                    .data_map
-                    .get::<XdgToplevelSurfaceData>()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .initial_configure_sent
-            });
-            if initial_configure_sent {
-                toplevel.send_configure();
-            }
-        }
-    }
-}
-delegate_xdg_decoration!(@<BEnd: Backend + 'static> Navda<BEnd>);
